adding thumbnail support

+++ MainWindow.xaml.cs
@@ -3,17 +3,9 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 using YoutubeExplode;
 using YoutubeExplode.Converter;
 using YoutubeExplode.Models;
@@ -24,9 +16,10 @@ namespace YouTubeWatcher
     public partial class MainWindow : Window
     {
         private IYoutubeClientHelper clientHelper;
-        private string webRootPath = "d:\\deleteme";
+        private string workingPath = "d:\\deleteme\\downloadedMedia";
         private string installLocation = System.AppDomain.CurrentDomain.BaseDirectory;
         private WebView wvMain;
+        private VideoDetails selectedVideoDetail;
 
         public MainWindow()
         {
@@ -62,22 +55,49 @@ namespace YouTubeWatcher
 
         private async Task GetVideoDetails() {
             if (!IsValidUrl()) return;
+            selectedVideoDetail = null;
             try {
-                var details = await clientHelper.GetVideoMetadata(clientHelper.GetVideoID(tbUrl.Text));
-                if (details != null)
+                selectedVideoDetail = await clientHelper.GetVideoMetadata(clientHelper.GetVideoID(tbUrl.Text));
+                if (selectedVideoDetail != null)
                 {
-                    foreach (var mt in details.qualities)
+                    foreach (var mt in selectedVideoDetail.qualities)
                     {
                         cbFormats.Items.Add(new ComboBoxItem() { Content = mt });
                     }
                 }
                 VideoChanged();
             }
-            catch (Exception ex) { 
+            catch (Exception ex) {
                 // todo: handle error
             }
-            
         }
+        bool isDownloadingThumb = false;
+        private async Task DownloadThumbnails() {
+            if (selectedVideoDetail == null) return;
+            if (isDownloadingThumb) return;
+            isDownloadingThumb = true;
+            await DownloadImageAsync($"{selectedVideoDetail.id}-low" , new Uri(selectedVideoDetail.thumbnails.LowResUrl));
+            await DownloadImageAsync($"{selectedVideoDetail.id}-medium", new Uri(selectedVideoDetail.thumbnails.MediumResUrl));
+            await DownloadImageAsync($"{selectedVideoDetail.id}-standard", new Uri(selectedVideoDetail.thumbnails.StandardResUrl));
+            await DownloadImageAsync($"{selectedVideoDetail.id}-high", new Uri(selectedVideoDetail.thumbnails.HighResUrl));
+            await DownloadImageAsync($"{selectedVideoDetail.id}-max", new Uri(selectedVideoDetail.thumbnails.MaxResUrl));
+            isDownloadingThumb = false;
+        }
+
+        private async Task DownloadImageAsync(string fileName, Uri uri)
+        {
+            using var httpClient = new System.Net.Http.HttpClient();
+
+            // Get the file extension
+            var uriWithoutQuery = uri.GetLeftPart(UriPartial.Path);
+            var fileExtension = System.IO.Path.GetExtension(uriWithoutQuery);
+
+            // Download the image and write to the file
+            var path = System.IO.Path.Combine(workingPath, $"{fileName}{fileExtension}");
+            var imageBytes = await httpClient.GetByteArrayAsync(uri);
+            await File.WriteAllBytesAsync(path, imageBytes);
+        }
+
         private bool IsValidUrl() {
             if (tbUrl.Text == "https://www.youtube.com/") return false;
             if (string.IsNullOrEmpty(tbUrl.Text)) return false;
@@ -86,31 +106,29 @@ namespace YouTubeWatcher
         private async void butLoad_Click(object sender, RoutedEventArgs e)
         {
             if (!IsValidUrl()) return;
+            if (selectedVideoDetail == null) return;
 
-            var details = await clientHelper.GetVideoMetadata(clientHelper.GetVideoID(tbUrl.Text));
-            var stream = new MediaStream();
+            await DownloadThumbnails();
+
             var mediaType = (string)((ComboBoxItem)cbMediaType.SelectedValue).Content;
             var quality = (string)((ComboBoxItem)cbFormats.SelectedValue).Content;
-            var mediaPath = webRootPath + $"\\DownloadedVideos\\{details.id}.{mediaType}";
+            var mediaPath = workingPath + $"\\{selectedVideoDetail.id}.{mediaType}";
 
             try
             {
                 if (File.Exists(mediaPath)) File.Delete(mediaPath);
-                await clientHelper.DownloadMedia(details.id, quality, mediaPath, mediaType);
+                await clientHelper.DownloadMedia(selectedVideoDetail.id, quality, mediaPath, mediaType);
             }
             catch (Exception ex)
             {
                 // todo: handle error
             }
 
-            var video = await clientHelper.GetVideoMetadata(details.id);
+            var stream = new MediaStream();
+            var video = await clientHelper.GetVideoMetadata(selectedVideoDetail.id);
             var videoStream = await stream.prepareMediaStream(mediaPath); // No need to dispose MemoryStream, GC will take care of this
             //CleanDirectory.DeleteFile(videoDir, id + ".mp4");
 
-            //if (videoStream == null)
-            //    return BadRequest();
-            //return File(videoStream, "video/mp4", video.Title);
-      
         }
     }
